requirements.txt
discord.py==1.7.1
python-dotenv~=0.17.0
SQLAlchemy~=1.4.7